Fixed Sin() output.
Changes:
	1. Changed the argument type of Sin() from int to double.

Bugs:
	Null!

+++ test/sin.h
@@ -4,8 +4,8 @@
 #define PRECISION 10
 
 double GetFactorial(int);	//	Returns the factorial of the provided number.
-double GetRadians(int);	//	Converts provide angle in degrees to radians. 
+double GetRadians(double);	//	Converts provide angle in degrees to radians. 
 
-double Sin(int);    //  Returns the sine of the provided angle in degrees, based on the Taylor series.
+double Sin(double);    //  Returns the sine of the provided angle in degrees, based on the Taylor series.
 
 #endif
